chuyendoiso/python-backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6 
pytesseract>=0.3.10
SQLAlchemy>=2.0.0
pyodbc>=5.0.1